% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{cor}
\alias{cor}
\alias{cor.default}
\alias{cor.dgCMatrix}
\title{Correlation (Sparse and Dense Matrices)}
\usage{
cor(x, ...)

\method{cor}{default}(x, ...)

\method{cor}{dgCMatrix}(x, ...)
}
\arguments{
\item{x}{A \code{matrix} or \code{dgCMatrix}.}

\item{...}{Additional arguments to pass to methods.}
}
\value{
A \code{p}x\code{p} matrix where \code{p} is the number of matrix columns, and the
(\code{i}, \code{j})th element corresponds to the sample correlation between \code{p_i}
and \code{p_j}.
}
\description{
\code{cor} computes the sample correlation between each column pair in a sparse
matrix.
}
\details{
\code{cor()} is an S3 generic with methods for:
\enumerate{
\item{
\code{dgCMatrix}
}
\item{
\code{matrix}
}
}
}
\examples{
x <- Matrix::rsparsematrix(10, 3, .5)
xdense <- as.matrix(x)

cor(x)
cor(xdense)

}
