% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-constant.R
\name{remove_constant}
\alias{remove_constant}
\alias{remove_constant.dgCMatrix}
\alias{remove_constant.matrix}
\title{Remove constant columns}
\usage{
remove_constant(x)

\method{remove_constant}{dgCMatrix}(x)

\method{remove_constant}{matrix}(x)
}
\arguments{
\item{x}{A \code{matrix} or \code{dgCMatrix}.}
}
\value{
\code{x} with zero-variance columns removed.
}
\description{
\code{remove_constant()} removes zero-variance columns from sparse and dense
matrices.
}
\details{
\code{remove_constant()} is an S3 generic with methods for:
\itemize{
\item{
\code{dgCMatrix}
}
\item{
\code{matrix}
}
}
}
\examples{
# Create a sparse matrix with constant columns
x <- Matrix::rsparsematrix(10, 3, 0.1)
colnames(x) <- paste0("x", 1:3)
x <- cbind(x, "x4" = 1, "x5" = 54)
# Print x
x

# Same matrix in dense format
xdense <- as.matrix(x)

# Drop constant columns
remove_constant(x)
remove_constant(xdense)

}
