% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-correlated.R
\name{remove_correlated}
\alias{remove_correlated}
\alias{remove_correlated.dgCMatrix}
\alias{remove_correlated.matrix}
\title{Remove highly correlated columns}
\usage{
remove_correlated(x, threshold)

\method{remove_correlated}{dgCMatrix}(x, threshold = 0.99)

\method{remove_correlated}{matrix}(x, threshold = 0.99)
}
\arguments{
\item{x}{A \code{matrix} or \code{dgCMatrix}.}

\item{threshold}{A double between 0 and 1 specifying the absolute correlation
threshold value at which to remove columns.}
}
\value{
\code{x} with one of highly correlated column-pairs removed.
}
\description{
\code{remove_correlated} removes one of column-pairs from sparse and dense
matrices that have sample correlation value greater than a user-defined
threshold.
}
\details{
\code{remove_correlated()} is an S3 generic with methods for:
\itemize{
\item{
\code{dgCMatrix}
}
\item{
\code{matrix}
}
}
}
\examples{
# Create a sparse matrix with very sparse columns
x <- Matrix::rsparsematrix(10, 5, 0.5)
x <- cbind(x, x[, 4:5], x[, 4:5])
# Create two perfectly correlated columns
colnames(x) <- paste0("x", 1:9)
# Print x
x

# Same matrix in dense format
xdense <- as.matrix(x)

# Drop highly correlated columns
remove_correlated(x, threshold = 0.99)
remove_correlated(xdense, threshold = 0.99)

}
