% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-duplicate.R
\name{remove_duplicate}
\alias{remove_duplicate}
\alias{remove_duplicate.dgCMatrix}
\alias{remove_duplicate.matrix}
\title{Remove duplicate columns}
\usage{
remove_duplicate(x)

\method{remove_duplicate}{dgCMatrix}(x)

\method{remove_duplicate}{matrix}(x)
}
\arguments{
\item{x}{A \code{matrix} or \code{dgCMatrix}.}
}
\value{
\code{x} with duplicate columns removed.
}
\description{
\code{remove_duplicate()} removes duplicate columns from sparse and dense
matrices.
}
\details{
\code{remove_duplicate()} is an S3 generic with methods for:
\enumerate{
\item{
\code{dgCMatrix}
}
\item{
\code{matrix}
}
}
}
\examples{
# Create a sparse matrix with duplicated columns
x <- Matrix::rsparsematrix(10, 4, 0.1)
colnames(x) <- paste0("x", 1:4)
x <- cbind(x, x[, c(1, 4)])
# Print x
x

# Same matrix in dense format
xdense <- as.matrix(x)

# Drop duplicate columns
remove_duplicate(x)
remove_duplicate(xdense)

}
