% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-empty.R
\name{remove_empty}
\alias{remove_empty}
\alias{remove_empty.CsparseMatrix}
\alias{remove_empty.matrix}
\title{Remove empty (all-zero) columns}
\usage{
remove_empty(x)

\method{remove_empty}{CsparseMatrix}(x)

\method{remove_empty}{matrix}(x)
}
\arguments{
\item{x}{A \code{matrix} or \code{CsparseMatrix}.}
}
\value{
\code{x} with all-zero columns removed.
}
\description{
\code{remove_empty()} removes all-zero columns from sparse and dense matrices.
}
\details{
\code{remove_empty()} is an S3 generic with methods for:
\itemize{
\item{
\code{CsparseMatrix}
}
\item{
\code{matrix}
}
}
}
\examples{
# Create a sparse matrix with constant columns
x <- Matrix::rsparsematrix(10, 3, 0.1)
x <- cbind(x, 0)
colnames(x) <- paste0("x", 1:4)
# Print x
x

# Same matrix in dense format
xdense <- as.matrix(x)

# Drop empty columns
remove_empty(x)
remove_empty(xdense)

}
