% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-sparse.R
\name{remove_sparse}
\alias{remove_sparse}
\alias{remove_sparse.dgCMatrix}
\alias{remove_sparse.matrix}
\title{Remove extremely sparse columns}
\usage{
remove_sparse(x, threshold)

\method{remove_sparse}{dgCMatrix}(x, threshold)

\method{remove_sparse}{matrix}(x, threshold)
}
\arguments{
\item{x}{A \code{matrix} or \code{dgCMatrix}.}

\item{threshold}{A double between 0 and 1 specifying the sparsity threshold
at which to remove columns.}
}
\value{
\code{x} with sparse columns removed.
}
\description{
\code{remove_sparse()} removes columns from sparse and dense matrices that have a
sparsity value greater than a user-defined threshold.
}
\details{
\code{remove_sparse()} is an S3 generic with methods for:
\itemize{
\item{
\code{dgCMatrix}
}
\item{
\code{matrix}
}
}
}
\examples{
# Create a sparse matrix with very sparse columns
x <- Matrix::rsparsematrix(10, 5, 0.1)
colnames(x) <- paste0("x", 1:5)
# Print x
x

# Same matrix in dense format
xdense <- as.matrix(x)

# Drop duplicate columns
remove_sparse(x, threshold = 0.9)
remove_sparse(xdense, threshold = 0.9)

}
