% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseprepr-package.R
\docType{package}
\name{sparseprepr-package}
\alias{sparseprepr}
\alias{sparseprepr-package}
\title{sparseprepr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Daniel Molitor \email{molitdj97@gmail.com}

}
\keyword{internal}
