% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseprepr-package.R
\docType{package}
\name{sparseprepr-package}
\alias{sparseprepr}
\alias{sparseprepr-package}
\title{sparseprepr: Pre Processing for Sparse Matrices}
\description{
Sparse matrices allow for in-memory modeling with very large data sets. This package enables common pre-processing actions for sparse matrices and provides a more memory-efficient workflow for modeling at scale.
}
\author{
\strong{Maintainer}: Daniel Molitor \email{molitdj97@gmail.com}

}
\keyword{internal}
